Separated geoutils.js from main code
